Title: 거리 합의 최솟값, Time: 14ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240520/거리 합의 최솟값/minimum-of-distance-sum.cpp
@@ -45,7 +45,7 @@ using namespace std;
 
 int main() {
     int N;
-    int answer = 0;
+    long long answer = 0;
     cin >> N;
 
     vector<int> rows(N);
